Allow for optimistic locking to be turned on post-production.
Data should be persisted in such a way that it's okay if optimistic
locking is turned on later, and shouldn't require a data migration.

This will require a migration for postgres, but will give a warning if
it hasn't been run. Existing data will need to be migrated - maybe delay
this for 2.0?